Add files via upload
Re-did both the pages. Had messed up the PHP section quite a lot so had to start over again.
I don't know why but the action page didn't work on my PC and the dropdown page started messing up when I deleted a few commented out snippets of the code. The drop down one still has issues but I guess it's better if the back end part is dealt with before we do anything to the front end.
Anyways lemme know if you get into any trouble with the design while coding.
Also, don't bother the icon, I was testing it a while ago. Not the final one obviously, it works but looks hideous.